feat: changed the groupId after passing the ownership to jumble-juggler organization
