+++ .github/workflows/build.yml
@@ -1,0 +1,38 @@
+name: Pull Request Build and Test
+
+on:
+  push:
+    branches:
+      - "**"
+    branches-ignore:
+      - master
+  pull_request:
+    branches:
+      - master
+
+jobs:
+  build-test-install:
+    permissions: write-all
+    runs-on: ubuntu-latest
+    steps:
+      - id: checkout-code
+        name: Checkout code
+        uses: actions/checkout@v2
+
+      - id: set-up-jdk
+        name: Set up JDK
+        uses: actions/setup-java@v4.0.0
+        with:
+          java-version: "17"
+          distribution: "adopt"
+          cache: "maven"
+
+      - id: mvn-verify
+        if: success()
+        name: Check if build is successful
+        run: mvn verify
+
+      - id: mvn-clean-install
+        if: success()
+        name: Build with Maven
+        run: mvn clean install
