+++ .github/workflows/release.yml
@@ -11,16 +11,17 @@ jobs:
     steps:
       - uses: actions/checkout@v3
       - name: Set up Java
-        uses: actions/setup-java@v3
+        uses: actions/setup-java@v4.0.0
         with:
-          java-version: '17'
-          distribution: 'adopt'
+          java-version: "17"
+          distribution: "adopt"
+          cache: "maven"
           server-id: ossrh # Value of the distributionManagement/repository/id field of the pom.xml
           server-username: OSSRH_USERNAME # env variable for username in deploy
           server-password: OSSRH_TOKEN # env variable for token in deploy
           gpg-private-key: ${{ secrets.OSSRH_GPG_SECRET_KEY }} # Value of the GPG private key to import
           gpg-passphrase: MAVEN_GPG_PASSPHRASE # env variable for GPG private key passphrase
-          
+
       - id: install-secret-key
         name: Install gpg secret key
         run: |
